Heavily refactored to add Search N Destroy.
BombArena = Sabotage
SndArena = Search N Destroy
EodArena = Explosives Ordnance Disposal

EodArena contains everything that is shared between Bomb & Snd.

This eliminates duplicate code.
And becomes easier to maintain because any changes to EodArena will
propagate down and effect BombArena & SndArena as well.

+++ src/main/java/mc/euro/demolition/holograms/HolographicDisplay.java
@@ -39,7 +39,10 @@ public class HolographicDisplay implements HologramInterface {
     
     @Override
     public void teleport(int id, Location loc) {
-        getHologram(id).teleport(loc.clone().add(0.0, 0.4, 0.0));
+        Hologram temp = getHologram(id);
+        if (temp != null) {
+            temp.teleport(loc.clone().add(0.0, 0.4, 0.0));
+        }         
     }
     
     @Override
@@ -56,10 +59,13 @@ public class HolographicDisplay implements HologramInterface {
     protected Hologram getHologram(int id) {
         return hmap.get(id);
     }
-    
+
     protected void deleteHologram(int id) {
-        Hologram hologram = hmap.get(id);
-        hologram.delete();
+        if (hmap.containsKey(id)) {
+            Hologram hologram = hmap.get(id);
+            hologram.delete();
+            hmap.remove(id);
+        }
     }
-    
+
 }
